+++ docs_src/configurations/staticfiles/example_multiple_fallthrough.py
@@ -3,8 +3,9 @@ from pathlib import Path
 from esmerald import Esmerald, StaticFilesConfig
 
 static_files_config = StaticFilesConfig(
-    path="/static", directory=["static/overwrites", "static", "static/defaults", "node_modules"],
-    name="static"
+    path="/static",
+    directory=["static/overwrites", "static", "static/defaults", "node_modules"],
+    name="static",
 )
 
 app = Esmerald(static_files_config=static_files_config)